MM-8 refactor: 토큰 유효 시간 단위 변경
