+++ src/main/java/com/dev/moim/domain/moim/dto/CalenderSchedule.java
@@ -1,4 +1,4 @@
-package com.dev.moim.domain.moim;
+package com.dev.moim.domain.moim.dto;
 
 import java.time.LocalDateTime;
