+++ src/main/java/com/dev/moim/domain/moim/dto/ParticipantProfile.java
@@ -1,4 +1,4 @@
-package com.dev.moim.domain.moim;
+package com.dev.moim.domain.moim.dto;
 
 public record ParticipantProfile(
         Long userId,
